// src/models/ICustomer.ts
export interface ICanopyEventSettings {
  overlays?: any[];
  meta?: any;
  show_live?: boolean;
  resizer_sizes?: number[];
  right_side_sizes?: number[];
}

export interface IUserCanopyAppsSettings {
  events?: Record<string, ICanopyEventSettings>;
  last_event_id?: string;
}

export interface IUserSiteAppsSettings {
  canopy?: IUserCanopyAppsSettings;
}

export interface IUserSiteSettings {
  theme?: string;
  apps?: IUserSiteAppsSettings;
}